BENCH: use intp dtype when indexing

+++ benchmarks/benchmarks/bench_ufunc.py
@@ -37,7 +37,7 @@ class Broadcast(Benchmark):
 class At(Benchmark):
     def setup(self):
         self.vals = np.random.rand(1_000_000)
-        self.idx = np.random.randint(1000, size=1_000_000)
+        self.idx = np.random.randint(1000, size=1_000_000).astype(np.intp)
         self.res = np.zeros(1000, dtype=self.vals.dtype)
 
     def time_sum_at(self):
